requirements.txt
qdrant-client
anthropic
fastembed
python-dotenv
numpy
